pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dataspoc-lens"
version = "0.1.0"
description = "Virtual warehouse — SQL over cloud Parquet via DuckDB"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.0",
    "typer>=0.9,<1.0",
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0",
    "fsspec>=2024.1",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "pygments>=2.0",
    "tabulate>=0.9",
]

[project.optional-dependencies]
s3 = ["s3fs>=2024.1"]
gcs = ["gcsfs>=2024.1"]
azure = ["adlfs>=2024.1"]
jupyter = ["jupyterlab>=4.0", "jupysql>=0.10", "duckdb-engine", "ipykernel", "numpy>=1.24", "pandas>=2.0"]
marimo = ["marimo>=0.6", "numpy>=1.24", "pandas>=2.0"]
ai = ["anthropic>=0.50", "openai>=1.0"]
all = ["dataspoc-lens[s3,gcs,azure,jupyter,marimo,ai]"]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "pyarrow>=15.0"]

[project.scripts]
dataspoc-lens = "dataspoc_lens.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/dataspoc_lens"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
